fix: isolate qr zip utils to client runtime
