+++ package.json
@@ -55,6 +55,8 @@
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
     "drizzle-orm": "^0.44.5",
+    "file-saver": "^2.0.5",
+    "jszip": "^3.10.1",
     "jose": "^6.1.0",
     "next": "14.2.5",
     "next-themes": "^0.3.0",
@@ -80,8 +82,6 @@
     "drizzle-kit": "^0.31.4",
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
-    "file-saver": "^2.0.5",
-    "jszip": "^3.10.1",
     "postcss": "^8",
     "qrcode": "^1.5.3",
     "readline-sync": "^1.4.10",
